✨ Refactor summarize button out
