+++ next/src/components/console/ChatWindow.tsx
@@ -11,8 +11,6 @@ import { FaArrowCircleDown } from "react-icons/fa";
 import { useAgentStore } from "../../stores";
 import { getTaskStatus, TASK_STATUS_EXECUTING } from "../../types/task";
 import { ImSpinner2 } from "react-icons/im";
-import Button from "../Button";
-import { useTaskStore } from "../../stores/taskStore";
 
 interface ChatWindowProps extends HeaderProps {
   children?: ReactNode;
@@ -91,7 +89,6 @@ const ChatWindow = ({
           );
         })}
         {children}
-        <Summarize />
         <div
           className={clsx(
             isThinking && !isStopped ? "opacity-100" : "opacity-0",
@@ -103,43 +100,6 @@ const ChatWindow = ({
           <ImSpinner2 className="animate-spin" />
         </div>
       </div>
-    </div>
-  );
-};
-
-const Summarize = () => {
-  const agent = useAgentStore.use.agent();
-  const lifecycle = useAgentStore.use.lifecycle();
-  const tasksWithResults = useTaskStore.use
-    .tasks()
-    .filter((task) => task.status == "completed" && task.result !== "");
-  const [summarized, setSummarized] = useState(false);
-
-  // Reset the summarized state when the agent changes
-  useEffect(() => {
-    setSummarized(false);
-  }, [agent]);
-
-  if (!agent || lifecycle !== "stopped" || tasksWithResults.length < 1 || summarized) return null;
-
-  return (
-    <div
-      className={clsx(
-        "mx-2 flex flex-row items-center gap-2 rounded-lg border border-white/20 p-2 font-mono transition duration-300 sm:mx-4",
-        "text-xs sm:text-base"
-      )}
-    >
-      <span className="md:hidden">Test</span>
-      <span className="hidden md:inline">Click here to summarize the conversation!</span>
-      <Button
-        className="ml-auto py-1  sm:py-1  md:py-1"
-        onClick={async () => {
-          setSummarized(true);
-          await agent?.summarize();
-        }}
-      >
-        Summarize
-      </Button>
     </div>
   );
 };
