+++ next/src/pages/index.tsx
@@ -34,6 +34,7 @@ import { ChatWindowTitle } from "../components/console/ChatWindowTitle";
 import { AgentApi } from "../services/agent/agent-api";
 import { toApiModelSettings } from "../utils/interfaces";
 import ExampleAgents from "../components/console/ExampleAgents";
+import Summarize from "../components/console/SummarizeButton";
 
 const Home: NextPage = () => {
   const { t } = useTranslation("indexPage");
@@ -188,6 +189,7 @@ const Home: NextPage = () => {
               visibleOnMobile={mobileVisibleWindow === "Chat"}
             >
               {messages.length === 0 && <ExampleAgents setAgentRun={setAgentRun} />}
+              <Summarize />
             </ChatWindow>
             <TaskWindow visibleOnMobile={mobileVisibleWindow === "Tasks"} />
           </Expand>
